// src/interfaces/ISignInTranslate.ts
export interface ISignInTranslate {
	signIn: string;
	signUp: string;
	phoneNumber: string;
	password: string;
	placeholderPhone: string;
	placeholderPassword: string;
	forgotpassword: string;
	forgotpassword2: string;
	validatePhone: string;
	validateOtp: string;
	pleaseFillOtpFromPhone: string;
	resendOtp: string;
	registerSuccess: string;
	noteForForgotPassword: string;
	receiveOtp: string;
	phoneAndPassWrong: string;
	resetPassword: string;
	registerNewPasswordSuccess: string;
	rulePassword: string;
	confirmPassword: string;
	passwordMatch: string;
	passwordNotMatch: string;
	bySelectingAgree: string;
	termsOfService: string;
	privacyPolicy: string;
	and: string;
	rePassword: string;
	signUpNow: string;
	sendOtp: string;
	pleaseFillOtp: string;
	submit: string;
	notReceivedOtp: string;
	createNewPassword: string;
	newPassword: string;
	reNewPassword: string;
	continue: string;
	phoneInvalid: string;
	requiredFillIn: string;
	otpInvalid: string;
	placeholderFillNewPassword: string;
	errorMatchPassword: string;
	otpIncorrect: string;
}